third commit

+++ task1.c
@@ -1,14 +1,19 @@
+// This program converts temperature from Celsius to Fahrenheit
+
+
 #include <stdio.h>
 
 int main() {
 
-int c, f;
-
+float c, f;
+// It prompts the user to enter a temperature in Celsius and then calculates the equivalent temperature in Fahrenheit using the formula: F = (C * 9/5) + 32
 printf("Enter temperature in Celsius: ");
 scanf("%d", &c);
 
+// Finally, it prints the result to the console.
 printf("Temperature in Fahrenheit: %d\n", f = (c * 9 / 5) + 32);
 
 return 0;
 }
+// This program converts temperature from Celsius to Fahrenheit
